Gère le cas ou l'agriculteur essaie d'ajouter un traitement phyto sans avoir de phyto dans son local

+++ src/AppBundle/Controller/ActionController.php
@@ -131,7 +131,7 @@ class ActionController extends Controller
             $em->flush();
 
             // Remove specialities added to wrong categories
-            if($action->getIntervention()->getInterventionCategory()->getSlug() != 'protection-des-cultures'){
+            if($action->getFarmSpecialityMvts() && $action->getIntervention()->getInterventionCategory()->getSlug() != 'protection-des-cultures'){
                 $mvts = $action->getFarmSpecialityMvts();
                 foreach($mvts as $mvt){
                     $em->remove($mvt);
